src/nat64: relocate inline tests to nat64_tests.rs
Same playbook. Extract 434-LOC inline test block from nat64.rs to
sibling nat64_tests.rs.

After: nat64.rs drops from 951 → 520 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
